+++ tests/test_unitcells.py
@@ -12,13 +12,13 @@ from conftest import (
     _gemmi_read_keys,
     additional_data_array,
     all_files_mark,
-    box_keys,
     cif_files_mark,
 )
 from gemmi import cif
 from more_itertools import flatten
 
 from parsnip import CifFile
+from parsnip._errors import ParseWarning
 
 
 def _gemmi_read_table(filename, keys):
@@ -28,39 +28,34 @@ def _gemmi_read_table(filename, keys):
         pytest.skip("Gemmi failed to read file!")
 
 
+@all_files_mark
+def test_read_symops(cif_data):
+    parsnip_symops = cif_data.file.symops
+    gemmi_symops = _gemmi_read_table(cif_data.filename, cif_data.file._symops_key)
+    np.testing.assert_array_equal(parsnip_symops, gemmi_symops)
+
+
 @all_files_mark  # TODO: test with conversions to numeric as well
 def test_read_wyckoff_positions(cif_data):
-    keys = ("_atom_site_fract_x", "_atom_site_fract_y", "_atom_site_fract_z")
     parsnip_data = cif_data.file.wyckoff_positions
-    gemmi_data = _gemmi_read_table(cif_data.filename, keys)
+    gemmi_data = _gemmi_read_table(cif_data.filename, cif_data.file._wyckoff_site_keys)
     gemmi_data = [[cif.as_number(val) for val in row] for row in gemmi_data]
     np.testing.assert_array_equal(parsnip_data, gemmi_data)
 
 
 @all_files_mark
 def test_read_cell_params(cif_data):
-    keys = box_keys
     parsnip_data = cif_data.file.read_cell_params()
-    gemmi_data = _gemmi_read_keys(cif_data.filename, keys)
+    cell_keys = cif_data.file._cell_keys
+    if cell_keys == []:
+        np.testing.assert_equal(parsnip_data, np.full((6,), np.nan))
+        return  # No cell keys found
+    gemmi_data = _gemmi_read_keys(cif_data.filename, cell_keys)
     np.testing.assert_array_equal(parsnip_data, gemmi_data)
 
     normalized = cif_data.file.read_cell_params(normalize=True)
     assert normalized[3:] == parsnip_data[3:]  # Should not change the angles
     assert min(normalized[:3]) == 1
-
-
-@all_files_mark
-def test_read_symmetry_operations(cif_data):
-    if "PDB_4INS_head.cif" in cif_data.filename:
-        return  # Excerpt of PDB file does not contain symmetry information
-
-    parsnip_data = cif_data.file.symops
-    gemmi_data = [
-        _gemmi_read_table(filename=cif_data.filename, keys=[k])
-        for k in cif_data.symop_keys
-    ]
-    gemmi_data = gemmi_data[0] if len(gemmi_data[0]) != 0 else gemmi_data[1]
-    np.testing.assert_array_equal(parsnip_data, gemmi_data)
 
 
 @cif_files_mark
@@ -137,6 +132,8 @@ def test_build_unit_cell(cif_data, n_decimal_places, parse_mode, cols):
 
     parsnip_minmax = [parsnip_positions.min(axis=0), parsnip_positions.max(axis=0)]
     ase_minmax = [ase_positions.min(axis=0), ase_positions.max(axis=0)]
+
+    np.testing.assert_array_equal(parsnip_positions.shape, ase_positions.shape)
     np.testing.assert_allclose(parsnip_minmax, ase_minmax, atol=1e-12)
 
     if cols is not None:
@@ -175,8 +172,10 @@ def test_invalid_unit_cell(cif_data):
 )
 @pytest.mark.parametrize("n_decimal_places", [3, 4])
 def test_build_accuracy(filename, n_decimal_places):
-    if "A5B10C8D4_mC108_15_a2ef_5f_4f_2f.cif" in filename or (
-        "A12B36CD12_cF488_210" in filename and n_decimal_places == 4
+    if (
+        "A5B10C8D4_mC108_15_a2ef_5f_4f_2f.cif" in filename
+        or "A2B2CD2_oP14_34_c_c_a_c.cif" in filename
+        or ("A12B36CD12_cF488_210" in filename and n_decimal_places == 4)
     ):
         pytest.xfail(reason="Known failing structure found.")
 
@@ -185,9 +184,11 @@ def test_build_accuracy(filename, n_decimal_places):
             return (False, -1)
         return (p.strip("'")[:2] == "hR", int(re.sub(r"[^\w]", "", p)[2:]))
 
+    warnings.filterwarnings("ignore", category=ParseWarning)
     cif = CifFile(filename)
+
     if cif["*Pearson"] is None:
-        pytest.skip(reason="Test not valid if Pearson symbol is unknown")
+        return
     (is_rhombohedral, n), uc = (
         parse_pearson(cif["*Pearson"]),
         cif.build_unit_cell(n_decimal_places=n_decimal_places, parse_mode="sympy"),